+++ bittytax/conv/mergers/etherscan.py
@@ -27,107 +27,170 @@ def merge_etherscan(data_files):
 
 def do_merge_etherscan(data_files, staking_addresses):
     merge = False
-
     tx_ids = {}
 
     for file_id in data_files:
-        if file_id not in (TOKENS, NFTS, INTERNAL_TXNS):
-            continue
-
         for dr in data_files[file_id].data_rows:
-            if dr.row_dict['Txhash'] in tx_ids:
-                tx_ids[dr.row_dict['Txhash']].append(MergeDataRow(dr, data_files[file_id], file_id))
-            else:
-                tx_ids[dr.row_dict['Txhash']] = [MergeDataRow(dr, data_files[file_id], file_id)]
+            if not dr.t_record:
+                continue
 
-    for data_row in data_files[TXNS].data_rows:
-        t_merge = [mdr for mdr in tx_ids.get(data_row.row_dict['Txhash'], [])
-                   if not mdr.data_row.parsed]
+            wallet = dr.t_record.wallet[- abs(TransactionOutRecord.WALLET_ADDR_LEN):]
+            if wallet not in tx_ids:
+                tx_ids[wallet] = {}
 
-        if t_merge:
-            if config.debug:
-                sys.stderr.write("%smerge: %s:%s\n" % (Fore.GREEN, TXNS.ljust(5), data_row))
+            if dr.row_dict['Txhash'] not in tx_ids[wallet]:
+                tx_ids[wallet][dr.row_dict['Txhash']] = []
 
-            for t in t_merge:
+            tx_ids[wallet][dr.row_dict['Txhash']]. \
+                    append(MergeDataRow(dr, data_files[file_id], file_id))
+
+    for wallet in tx_ids:
+        for txn in tx_ids[wallet]:
+            if len(tx_ids[wallet][txn]) == 1:
                 if config.debug:
-                    sys.stderr.write("%smerge: %s:%s\n" % (Fore.GREEN, t.data_file_id.ljust(5),
-                                                           t.data_row))
+                    sys.stderr.write("%smerge: %s:%s\n" % (
+                        Fore.BLUE,
+                        tx_ids[wallet][txn][0].data_file_id.ljust(5),
+                        tx_ids[wallet][txn][0].data_row))
+                continue
 
-                data_row.parsed = True
-        else:
+            for t in tx_ids[wallet][txn]:
+                if config.debug:
+                    sys.stderr.write("%smerge: %s:%s\n" % (
+                        Fore.GREEN,
+                        t.data_file_id.ljust(5),
+                        t.data_row))
+
+            t_ins, t_outs, t_fee = get_ins_outs(tx_ids[wallet][txn])
+
             if config.debug:
-                sys.stderr.write("%smerge: txn:  %s\n" % (Fore.BLUE, data_row))
+                output_records(t_ins, t_outs, t_fee)
+                sys.stderr.write("%smerge:     consolidate:\n" % (Fore.YELLOW))
 
-            continue
+            consolidate(tx_ids[wallet][txn], [TXNS, INTERNAL_TXNS])
 
-        t_ins = [t.data_row for t in t_merge if t.data_row.t_record and
-                 t.data_row.t_record.t_type == TransactionOutRecord.TYPE_DEPOSIT]
-        t_outs = [t.data_row for t in t_merge if t.data_row.t_record and
-                  t.data_row.t_record.t_type == TransactionOutRecord.TYPE_WITHDRAWAL]
+            t_ins, t_outs, t_fee = get_ins_outs(tx_ids[wallet][txn])
 
-        if data_row.t_record.t_type == TransactionOutRecord.TYPE_DEPOSIT:
-            t_ins.append(data_row)
-        elif data_row.t_record.t_type == TransactionOutRecord.TYPE_WITHDRAWAL:
-            t_outs.append(data_row)
+            if config.debug:
+                output_records(t_ins, t_outs, t_fee)
+                sys.stderr.write("%smerge:     merge:\n" % (Fore.YELLOW))
 
-        if config.debug:
-            output_records(data_row, t_ins, t_outs)
+            if t_fee:
+                fee_quantity = t_fee.t_record.fee_quantity
+                fee_asset = t_fee.t_record.fee_asset
 
-        t_ins_orig = copy.copy(t_ins)
-        fee_quantity = data_row.t_record.fee_quantity
-        fee_asset = data_row.t_record.fee_asset
+            t_ins_orig = copy.copy(t_ins)
+            if t_fee:
+                method_handling(t_ins, t_fee, staking_addresses)
 
-        method_handling(t_ins, data_row, staking_addresses)
+            # Make trades
+            if len(t_ins) == 1 and t_outs:
+                do_etherscan_multi_sell(t_ins, t_outs, t_fee)
+            elif len(t_outs) == 1 and t_ins:
+                do_etherscan_multi_buy(t_ins, t_outs, t_fee)
+            elif len(t_ins) > 1 and len(t_outs) > 1:
+                # multi-sell to multi-buy trade not supported
+                sys.stderr.write("%sWARNING%s Merge failure for Txhash: %s\n" % (
+                    Back.YELLOW+Fore.BLACK, Back.RESET+Fore.YELLOW, txn))
 
-        # Make trades
-        if len(t_ins) == 1 and t_outs:
-            do_etherscan_multi_sell(t_ins, t_outs, data_row)
-        elif len(t_outs) == 1 and t_ins:
-            do_etherscan_multi_buy(t_ins, t_outs, data_row)
-        elif len(t_ins) > 1 and len(t_outs) > 1:
-            # multi-sell to multi-buy trade not supported
-            sys.stderr.write("%sWARNING%s Merge failure for Txhash: %s\n" % (
-                Back.YELLOW+Fore.BLACK, Back.RESET+Fore.YELLOW, data_row.row_dict['Txhash']))
+                for mdr in tx_ids[wallet][txn]:
+                    mdr.data_row.failure = UnexpectedContentError(
+                        mdr.data_file.parser.in_header.index('Txhash'),
+                        'Txhash', mdr.data_row.row_dict['Txhash'])
+                    sys.stderr.write("%srow[%s] %s\n" % (
+                        Fore.YELLOW,
+                        mdr.data_file.parser.in_header_row_num + mdr.data_row.line_num,
+                        mdr.data_row))
+                continue
 
-            for mdr in t_merge:
-                mdr.data_row.failure = UnexpectedContentError(
-                    mdr.data_file.parser.in_header.index('Txhash'),
-                    'Txhash', mdr.data_row.row_dict['Txhash'])
-                sys.stderr.write("%srow[%s] %s\n" % (
-                    Fore.YELLOW,
-                    mdr.data_file.parser.in_header_row_num + mdr.data_row.line_num,
-                    mdr.data_row))
-            continue
+            if t_fee:
+                # Split fees
+                t_all = [t for t in t_ins_orig + t_outs if t.t_record]
+                do_fee_split(t_all, t_fee, fee_quantity, fee_asset)
 
-        # Split fees
-        t_all = [t for t in t_ins_orig + t_outs if t.t_record]
-        do_fee_split(t_all, data_row, fee_quantity, fee_asset)
+            merge = True
 
-        merge = True
-
-        if config.debug:
-            output_records(data_row, t_ins_orig, t_outs)
+            if config.debug:
+                output_records(t_ins_orig, t_outs, t_fee)
 
     return merge
 
-def output_records(data_row, t_ins, t_outs):
-    dup = bool(data_row in t_ins + t_outs)
+def get_ins_outs(tx_ids):
+    t_ins = [t.data_row for t in tx_ids if t.data_row.t_record and
+             t.data_row.t_record.t_type == TransactionOutRecord.TYPE_DEPOSIT]
+    t_outs = [t.data_row for t in tx_ids if t.data_row.t_record and
+              t.data_row.t_record.t_type == TransactionOutRecord.TYPE_WITHDRAWAL]
+    t_fees = [t.data_row for t in tx_ids if t.data_row.t_record and
+              t.data_row.t_record.fee_quantity]
 
-    sys.stderr.write("%smerge:   TR-F%s: %s\n" % (
-        Fore.YELLOW, '*' if dup else '', data_row.t_record))
+    if len(t_fees) == 0:
+        t_fee = None
+    elif len(t_fees) == 1:
+        t_fee = t_fees[0]
+    else:
+        raise Exception
+
+    return t_ins, t_outs, t_fee
+
+def consolidate(tx_ids, file_ids):
+    tx_assets = {}
+
+    for txn in list(tx_ids):
+        if txn.data_file_id not in file_ids:
+            return
+
+        asset = txn.data_row.t_record.get_asset()
+        if asset not in tx_assets:
+            tx_assets[asset] = txn
+            tx_assets[asset].quantity += txn.data_row.t_record.get_quantity()
+        else:
+            tx_assets[asset].quantity += txn.data_row.t_record.get_quantity()
+            txn.data_row.t_record = None
+            tx_ids.remove(txn)
+
+    for asset in tx_assets:
+        txn = tx_assets[asset]
+        if txn.quantity > 0:
+            txn.data_row.t_record.t_type = TransactionOutRecord.TYPE_DEPOSIT
+            txn.data_row.t_record.buy_asset = asset
+            txn.data_row.t_record.buy_quantity = txn.quantity
+            txn.data_row.t_record.sell_asset = ''
+            txn.data_row.t_record.sell_quantity = None
+        elif tx_assets[asset].quantity < 0:
+            txn.data_row.t_record.t_type = TransactionOutRecord.TYPE_WITHDRAWAL
+            txn.data_row.t_record.buy_asset = ''
+            txn.data_row.t_record.buy_quantity = None
+            txn.data_row.t_record.sell_asset = asset
+            txn.data_row.t_record.sell_quantity = abs(txn.quantity)
+        else:
+            if txn.data_row.t_record.fee_quantity:
+                txn.data_row.t_record.t_type = TransactionOutRecord.TYPE_SPEND
+                txn.data_row.t_record.buy_asset = ''
+                txn.data_row.t_record.buy_quantity = None
+                txn.data_row.t_record.sell_asset = asset
+                txn.data_row.t_record.sell_quantity = Decimal(0)
+            else:
+                tx_ids.remove(txn)
+
+def output_records(t_ins, t_outs, t_fee):
+    dup = bool(t_fee and t_fee in t_ins + t_outs)
+
+    if t_fee:
+        sys.stderr.write("%smerge:   TR-F%s: %s\n" % (
+            Fore.YELLOW, '*' if dup else '', t_fee.t_record))
 
     for t_in in t_ins:
         sys.stderr.write("%smerge:   TR-I%s: %s\n" % (
-            Fore.YELLOW, '*' if data_row is t_in else '', t_in.t_record))
+            Fore.YELLOW, '*' if t_fee is t_in else '', t_in.t_record))
     for t_out in t_outs:
         sys.stderr.write("%smerge:   TR-O%s: %s\n" % (
-            Fore.YELLOW, '*' if data_row is t_out else '', t_out.t_record))
+            Fore.YELLOW, '*' if t_fee is t_out else '', t_out.t_record))
 
-def method_handling(t_ins, data_row, staking_addresses):
-    if data_row.row_dict.get('Method') in ("Enter Staking", "Leave Staking", "Deposit", "Withdraw"):
+def method_handling(t_ins, t_fee, staking_addresses):
+    if t_fee.row_dict.get('Method') in ("Enter Staking", "Leave Staking", "Deposit", "Withdraw"):
         if t_ins:
             staking = [t for t in t_ins if t.row_dict['ContractAddress'] in staking_addresses
-                       and t.row_dict['From'] != data_row.row_dict['To']]
+                       and t.row_dict['From'] != t_fee.row_dict['To']]
             if staking:
                 if len(staking) == 1:
                     staking[0].t_record.t_type = TransactionOutRecord.TYPE_STAKING
@@ -138,7 +201,7 @@ def method_handling(t_ins, data_row, staking_addresses):
                 else:
                     raise Exception
 
-def do_etherscan_multi_sell(t_ins, t_outs, data_row):
+def do_etherscan_multi_sell(t_ins, t_outs, t_fee):
     if config.debug:
         sys.stderr.write("%smerge:     trade sell(s):\n" % (Fore.YELLOW))
 
@@ -168,12 +231,13 @@ def do_etherscan_multi_sell(t_ins, t_outs, data_row):
         t_out.t_record.t_type = TransactionOutRecord.TYPE_TRADE
         t_out.t_record.buy_quantity = split_buy_quantity
         t_out.t_record.buy_asset = buy_asset
-        t_out.t_record.note = get_note(data_row.row_dict)
+        if t_fee:
+            t_out.t_record.note = get_note(t_fee.row_dict)
 
     # Remove TR for buy now it's been added to each sell
     t_ins[0].t_record = None
 
-def do_etherscan_multi_buy(t_ins, t_outs, data_row):
+def do_etherscan_multi_buy(t_ins, t_outs, t_fee):
     if config.debug:
         sys.stderr.write("%smerge:     trade buy(s):\n" % (Fore.YELLOW))
 
@@ -203,12 +267,13 @@ def do_etherscan_multi_buy(t_ins, t_outs, data_row):
         t_in.t_record.t_type = TransactionOutRecord.TYPE_TRADE
         t_in.t_record.sell_quantity = split_sell_quantity
         t_in.t_record.sell_asset = sell_asset
-        t_in.t_record.note = get_note(data_row.row_dict)
+        if t_fee:
+            t_in.t_record.note = get_note(t_fee.row_dict)
 
     # Remove TR for sell now it's been added to each buy
     t_outs[0].t_record = None
 
-def do_fee_split(t_all, data_row, fee_quantity, fee_asset):
+def do_fee_split(t_all, t_fee, fee_quantity, fee_asset):
     if config.debug:
         sys.stderr.write("%smerge:     split fees:\n" % (Fore.YELLOW))
         sys.stderr.write("%smerge:       fee_quantity=%s fee_asset=%s\n" % (
@@ -232,15 +297,15 @@ def do_fee_split(t_all, data_row, fee_quantity, fee_asset):
 
         t.t_record.fee_quantity = split_fee_quantity
         t.t_record.fee_asset = fee_asset
-        t.t_record.note = get_note(data_row.row_dict)
+        t.t_record.note = get_note(t_fee.row_dict)
 
     # Remove TR for fee now it's been added to each withdrawal
-    if data_row.t_record and data_row not in t_all:
-        if data_row.t_record.t_type == TransactionOutRecord.TYPE_SPEND:
-            data_row.t_record = None
+    if t_fee.t_record and t_fee not in t_all:
+        if t_fee.t_record.t_type == TransactionOutRecord.TYPE_SPEND:
+            t_fee.t_record = None
         else:
-            data_row.t_record.fee_quantity = None
-            data_row.t_record.fee_asset = ''
+            t_fee.t_record.fee_quantity = None
+            t_fee.t_record.fee_asset = ''
 
 DataMerge("Etherscan fees & multi-token transactions",
           {TXNS: {'req': DataMerge.MAN, 'obj': etherscan_txns},