+++ bittytax/conv/mergers/snowtrace.py
@@ -3,7 +3,8 @@
 
 from .etherscan import do_merge_etherscan, TXNS, TOKENS, NFTS, INTERNAL_TXNS
 from ..datamerge import DataMerge
-from ..parsers.snowtrace import avax_txns, avax_int, WALLET
+from ..out_record import TransactionOutRecord
+from ..parsers.snowtrace import avax_txns, avax_int, WALLET, WORKSHEET_NAME
 from ..parsers.etherscan import etherscan_tokens, etherscan_nfts
 
 STAKE_ADDRESSES = []
@@ -15,16 +16,18 @@ def merge_snowtrace(data_files):
     if merge:
         # Change Etherscan parsers to SnowTrace
         if TOKENS in data_files:
-            data_files[TOKENS].parser.worksheet_name = "SnowTrace"
+            data_files[TOKENS].parser.worksheet_name = WORKSHEET_NAME
             for data_row in data_files[TOKENS].data_rows:
                 if data_row.t_record:
-                    data_row.t_record.wallet = WALLET
+                    address = data_row.t_record.wallet[- abs(TransactionOutRecord.WALLET_ADDR_LEN):]
+                    data_row.t_record.wallet = "%s-%s" % (WALLET, address)
 
         if NFTS in data_files:
-            data_files[NFTS].parser.worksheet_name = "SnowTrace"
+            data_files[NFTS].parser.worksheet_name = WORKSHEET_NAME
             for data_row in data_files[NFTS].data_rows:
                 if data_row.t_record:
-                    data_row.t_record.wallet = WALLET
+                    address = data_row.t_record.wallet[- abs(TransactionOutRecord.WALLET_ADDR_LEN):]
+                    data_row.t_record.wallet = "%s-%s" % (WALLET, address)
 
     return merge