+++ bittytax/conv/parsers/bscscan.py
@@ -8,6 +8,7 @@ from ..out_record import TransactionOutRecord
 from ..dataparser import DataParser
 
 WALLET = "Binance Smart Chain"
+WORKSHEET_NAME = "BscScan"
 
 def parse_bscscan(data_row, _parser, **_kwargs):
     row_dict = data_row.row_dict
@@ -23,7 +24,7 @@ def parse_bscscan(data_row, _parser, **_kwargs):
                                                      data_row.timestamp,
                                                      buy_quantity=row_dict['Value_IN(BNB)'],
                                                      buy_asset="BNB",
-                                                     wallet=WALLET,
+                                                     wallet=get_wallet(row_dict['To']),
                                                      note=get_note(row_dict))
     elif Decimal(row_dict['Value_OUT(BNB)']) > 0:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_WITHDRAWAL,
@@ -32,7 +33,7 @@ def parse_bscscan(data_row, _parser, **_kwargs):
                                                  sell_asset="BNB",
                                                  fee_quantity=row_dict['TxnFee(BNB)'],
                                                  fee_asset="BNB",
-                                                 wallet=WALLET,
+                                                 wallet=get_wallet(row_dict['From']),
                                                  note=get_note(row_dict))
     else:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_SPEND,
@@ -41,25 +42,32 @@ def parse_bscscan(data_row, _parser, **_kwargs):
                                                  sell_asset="BNB",
                                                  fee_quantity=row_dict['TxnFee(BNB)'],
                                                  fee_asset="BNB",
-                                                 wallet=WALLET,
+                                                 wallet=get_wallet(row_dict['From']),
                                                  note=get_note(row_dict))
+
+def get_wallet(address):
+    return "%s-%s" % (WALLET, address.lower()[0:TransactionOutRecord.WALLET_ADDR_LEN])
 
 def parse_bscscan_internal(data_row, _parser, **_kwargs):
     row_dict = data_row.row_dict
     data_row.timestamp = DataParser.parse_timestamp(int(row_dict['UnixTimestamp']))
+
+    # Failed internal txn
+    if row_dict['Status'] != '0':
+        return
 
     if Decimal(row_dict['Value_IN(BNB)']) > 0:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_DEPOSIT,
                                                  data_row.timestamp,
                                                  buy_quantity=row_dict['Value_IN(BNB)'],
                                                  buy_asset="BNB",
-                                                 wallet=WALLET)
+                                                 wallet=get_wallet(row_dict['TxTo']))
     elif Decimal(row_dict['Value_OUT(BNB)']) > 0:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_WITHDRAWAL,
                                                  data_row.timestamp,
                                                  sell_quantity=row_dict['Value_OUT(BNB)'],
                                                  sell_asset="BNB",
-                                                 wallet=WALLET)
+                                                 wallet=get_wallet(row_dict['From']))
 
 bsc_txns = DataParser(
         DataParser.TYPE_EXPLORER,
@@ -67,7 +75,7 @@ bsc_txns = DataParser(
         ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
          'Value_IN(BNB)', 'Value_OUT(BNB)', None, 'TxnFee(BNB)', 'TxnFee(USD)',
          'Historical $Price/BNB', 'Status', 'ErrCode'],
-        worksheet_name="BscScan",
+        worksheet_name=WORKSHEET_NAME,
         row_handler=parse_bscscan)
 
 DataParser(DataParser.TYPE_EXPLORER,
@@ -75,7 +83,7 @@ DataParser(DataParser.TYPE_EXPLORER,
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
             'Value_IN(BNB)', 'Value_OUT(BNB)', None, 'TxnFee(BNB)', 'TxnFee(USD)',
             'Historical $Price/BNB', 'Status', 'ErrCode', 'PrivateNote'],
-           worksheet_name="BscScan",
+           worksheet_name=WORKSHEET_NAME,
            row_handler=parse_bscscan)
 
 DataParser(DataParser.TYPE_EXPLORER,
@@ -83,7 +91,7 @@ DataParser(DataParser.TYPE_EXPLORER,
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
             'Value_IN(BNB)', 'Value_OUT(BNB)', None, 'TxnFee(BNB)', 'TxnFee(USD)',
             'Historical $Price/BNB', 'Status', 'ErrCode', 'Method'],
-           worksheet_name="BscScan",
+           worksheet_name=WORKSHEET_NAME,
            row_handler=parse_bscscan)
 
 DataParser(DataParser.TYPE_EXPLORER,
@@ -91,7 +99,7 @@ DataParser(DataParser.TYPE_EXPLORER,
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
             'Value_IN(BNB)', 'Value_OUT(BNB)', None, 'TxnFee(BNB)', 'TxnFee(USD)',
             'Historical $Price/BNB', 'Status', 'ErrCode', 'Method', 'PrivateNote'],
-           worksheet_name="BscScan",
+           worksheet_name=WORKSHEET_NAME,
            row_handler=parse_bscscan)
 
 bsc_int = DataParser(
@@ -100,7 +108,7 @@ bsc_int = DataParser(
         ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'ParentTxFrom', 'ParentTxTo',
          'ParentTxETH_Value', 'From', 'TxTo', 'ContractAddress', 'Value_IN(BNB)',
          'Value_OUT(BNB)', None, 'Historical $Price/BNB', 'Status', 'ErrCode', 'Type'],
-        worksheet_name="BscScan",
+        worksheet_name=WORKSHEET_NAME,
         row_handler=parse_bscscan_internal)
 
 DataParser(DataParser.TYPE_EXPLORER,
@@ -109,7 +117,7 @@ DataParser(DataParser.TYPE_EXPLORER,
             'ParentTxETH_Value', 'From', 'TxTo', 'ContractAddress', 'Value_IN(BNB)',
             'Value_OUT(BNB)', None, 'Historical $Price/BNB', 'Status', 'ErrCode', 'Type',
             'PrivateNote'],
-           worksheet_name="BscScan",
+           worksheet_name=WORKSHEET_NAME,
            row_handler=parse_bscscan_internal)
 
 # Same header as Etherscan
@@ -117,7 +125,7 @@ DataParser(DataParser.TYPE_EXPLORER,
 #           "BscScan (BEP-20 Tokens)",
 #           ['Txhash', 'UnixTimestamp', 'DateTime', 'From', 'To', 'Value', 'ContractAddress',
 #            'TokenName', 'TokenSymbol'],
-#           worksheet_name="BscScan",
+#           worksheet_name=WORKSHEET_NAME,
 #           row_handler=parse_bscscan_tokens)
 
 # Same header as Etherscan
@@ -125,5 +133,5 @@ DataParser(DataParser.TYPE_EXPLORER,
 #           "Etherscan (BEP-721 NFTs)",
 #           ['Txhash', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress', 'TokenId',
 #            'TokenName', 'TokenSymbol'],
-#           worksheet_name="BscScan",
+#           worksheet_name=WORKSHEET_NAME,
 #           row_handler=parse_bscscan_nfts)