+++ bittytax/conv/parsers/etherscan.py
@@ -23,7 +23,7 @@ def parse_etherscan(data_row, _parser, **_kwargs):
                                                      data_row.timestamp,
                                                      buy_quantity=row_dict['Value_IN(ETH)'],
                                                      buy_asset="ETH",
-                                                     wallet=WALLET,
+                                                     wallet=get_wallet(row_dict['To']),
                                                      note=get_note(row_dict))
     elif Decimal(row_dict['Value_OUT(ETH)']) > 0:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_WITHDRAWAL,
@@ -32,7 +32,7 @@ def parse_etherscan(data_row, _parser, **_kwargs):
                                                  sell_asset="ETH",
                                                  fee_quantity=row_dict['TxnFee(ETH)'],
                                                  fee_asset="ETH",
-                                                 wallet=WALLET,
+                                                 wallet=get_wallet(row_dict['From']),
                                                  note=get_note(row_dict))
     else:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_SPEND,
@@ -41,8 +41,11 @@ def parse_etherscan(data_row, _parser, **_kwargs):
                                                  sell_asset="ETH",
                                                  fee_quantity=row_dict['TxnFee(ETH)'],
                                                  fee_asset="ETH",
-                                                 wallet=WALLET,
+                                                 wallet=get_wallet(row_dict['From']),
                                                  note=get_note(row_dict))
+
+def get_wallet(address):
+    return "%s-%s" % (WALLET, address.lower()[0:TransactionOutRecord.WALLET_ADDR_LEN])
 
 def get_note(row_dict):
     if row_dict['Status'] != '':
@@ -59,18 +62,22 @@ def parse_etherscan_internal(data_row, _parser, **_kwargs):
     row_dict = data_row.row_dict
     data_row.timestamp = DataParser.parse_timestamp(int(row_dict['UnixTimestamp']))
 
+    # Failed internal txn
+    if row_dict['Status'] != '0':
+        return
+
     if Decimal(row_dict['Value_IN(ETH)']) > 0:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_DEPOSIT,
                                                  data_row.timestamp,
                                                  buy_quantity=row_dict['Value_IN(ETH)'],
                                                  buy_asset="ETH",
-                                                 wallet=WALLET)
+                                                 wallet=get_wallet(row_dict['TxTo']))
     elif Decimal(row_dict['Value_OUT(ETH)']) > 0:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_WITHDRAWAL,
                                                  data_row.timestamp,
                                                  sell_quantity=row_dict['Value_OUT(ETH)'],
                                                  sell_asset="ETH",
-                                                 wallet=WALLET)
+                                                 wallet=get_wallet(row_dict['From']))
 
 def parse_etherscan_tokens(data_row, _parser, **kwargs):
     row_dict = data_row.row_dict
@@ -86,13 +93,13 @@ def parse_etherscan_tokens(data_row, _parser, **kwargs):
                                                  data_row.timestamp,
                                                  buy_quantity=row_dict['Value'].replace(',', ''),
                                                  buy_asset=asset,
-                                                 wallet=WALLET)
+                                                 wallet=get_wallet(row_dict['To']))
     elif row_dict['From'].lower() in kwargs['filename'].lower():
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_WITHDRAWAL,
                                                  data_row.timestamp,
                                                  sell_quantity=row_dict['Value'].replace(',', ''),
                                                  sell_asset=asset,
-                                                 wallet=WALLET)
+                                                 wallet=get_wallet(row_dict['From']))
     else:
         raise DataFilenameError(kwargs['filename'], "Ethereum address")
 
@@ -106,14 +113,14 @@ def parse_etherscan_nfts(data_row, _parser, **kwargs):
                                                  buy_quantity=1,
                                                  buy_asset='{} #{}'.format(row_dict['TokenName'],
                                                                            row_dict['TokenId']),
-                                                 wallet=WALLET)
+                                                 wallet=get_wallet(row_dict['To']))
     elif row_dict['From'].lower() in kwargs['filename'].lower():
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_WITHDRAWAL,
                                                  data_row.timestamp,
                                                  sell_quantity=1,
                                                  sell_asset='{} #{}'.format(row_dict['TokenName'],
                                                                             row_dict['TokenId']),
-                                                 wallet=WALLET)
+                                                 wallet=get_wallet(row_dict['From']))
     else:
         raise DataFilenameError(kwargs['filename'], "Ethereum address")