+++ bittytax/conv/parsers/hecoinfo.py
@@ -8,6 +8,7 @@ from ..out_record import TransactionOutRecord
 from ..dataparser import DataParser
 
 WALLET = "Huobi Eco Chain"
+WORKSHEET_NAME = "HecoInfo"
 
 def parse_hecoinfo(data_row, _parser, **_kwargs):
     row_dict = data_row.row_dict
@@ -23,7 +24,7 @@ def parse_hecoinfo(data_row, _parser, **_kwargs):
                                                      data_row.timestamp,
                                                      buy_quantity=row_dict['Value_IN(HT)'],
                                                      buy_asset="HT",
-                                                     wallet=WALLET,
+                                                     wallet=get_wallet(row_dict['To']),
                                                      note=get_note(row_dict))
 
     elif Decimal(row_dict['Value_OUT(HT)']) > 0:
@@ -33,7 +34,7 @@ def parse_hecoinfo(data_row, _parser, **_kwargs):
                                                  sell_asset="HT",
                                                  fee_quantity=row_dict['TxnFee(HT)'],
                                                  fee_asset="HT",
-                                                 wallet=WALLET,
+                                                 wallet=get_wallet(row_dict['From']),
                                                  note=get_note(row_dict))
     else:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_SPEND,
@@ -42,25 +43,32 @@ def parse_hecoinfo(data_row, _parser, **_kwargs):
                                                  sell_asset="HT",
                                                  fee_quantity=row_dict['TxnFee(HT)'],
                                                  fee_asset="HT",
-                                                 wallet=WALLET,
+                                                 wallet=get_wallet(row_dict['From']),
                                                  note=get_note(row_dict))
+
+def get_wallet(address):
+    return "%s-%s" % (WALLET, address.lower()[0:TransactionOutRecord.WALLET_ADDR_LEN])
 
 def parse_hecoinfo_internal(data_row, _parser, **_kwargs):
     row_dict = data_row.row_dict
     data_row.timestamp = DataParser.parse_timestamp(int(row_dict['UnixTimestamp']))
+
+    # Failed internal txn
+    if row_dict['Status'] != '0':
+        return
 
     if Decimal(row_dict['Value_IN(HT)']) > 0:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_DEPOSIT,
                                                  data_row.timestamp,
                                                  buy_quantity=row_dict['Value_IN(HT)'],
                                                  buy_asset="HT",
-                                                 wallet=WALLET)
+                                                 wallet=get_wallet(row_dict['TxTo']))
     elif Decimal(row_dict['Value_OUT(HT)']) > 0:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_WITHDRAWAL,
                                                  data_row.timestamp,
                                                  sell_quantity=row_dict['Value_OUT(HT)'],
                                                  sell_asset="HT",
-                                                 wallet=WALLET)
+                                                 wallet=get_wallet(row_dict['From']))
 
 heco_txns = DataParser(
         DataParser.TYPE_EXPLORER,
@@ -68,7 +76,7 @@ heco_txns = DataParser(
         ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
          'Value_IN(HT)', 'Value_OUT(HT)', None, 'TxnFee(HT)', 'TxnFee(USD)',
          'Historical $Price/HT', 'Status', 'ErrCode'],
-        worksheet_name="HecoInfo",
+        worksheet_name=WORKSHEET_NAME,
         row_handler=parse_hecoinfo)
 
 DataParser(DataParser.TYPE_EXPLORER,
@@ -76,7 +84,7 @@ DataParser(DataParser.TYPE_EXPLORER,
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
             'Value_IN(HT)', 'Value_OUT(HT)', None, 'TxnFee(HT)', 'TxnFee(USD)',
             'Historical $Price/HT', 'Status', 'ErrCode', 'PrivateNote'],
-           worksheet_name="HecoInfo",
+           worksheet_name=WORKSHEET_NAME,
            row_handler=parse_hecoinfo)
 
 DataParser(DataParser.TYPE_EXPLORER,
@@ -84,7 +92,7 @@ DataParser(DataParser.TYPE_EXPLORER,
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
             'Value_IN(HT)', 'Value_OUT(HT)', None, 'TxnFee(HT)', 'TxnFee(USD)',
             'Historical $Price/HT', 'Status', 'ErrCode', 'Method'],
-           worksheet_name="HecoInfo",
+           worksheet_name=WORKSHEET_NAME,
            row_handler=parse_hecoinfo)
 
 DataParser(DataParser.TYPE_EXPLORER,
@@ -92,7 +100,7 @@ DataParser(DataParser.TYPE_EXPLORER,
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
             'Value_IN(HT)', 'Value_OUT(HT)', None, 'TxnFee(HT)', 'TxnFee(USD)',
             'Historical $Price/HT', 'Status', 'ErrCode', 'Method', 'PrivateNote'],
-           worksheet_name="HecoInfo",
+           worksheet_name=WORKSHEET_NAME,
            row_handler=parse_hecoinfo)
 
 heco_int = DataParser(
@@ -101,7 +109,7 @@ heco_int = DataParser(
         ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'ParentTxFrom', 'ParentTxTo',
          'ParentTxETH_Value', 'From', 'TxTo', 'ContractAddress', 'Value_IN(HT)', 'Value_OUT(HT)',
          None, 'Historical $Price/HT', 'Status', 'ErrCode', 'Type'],
-        worksheet_name="HecoInfo",
+        worksheet_name=WORKSHEET_NAME,
         row_handler=parse_hecoinfo_internal)
 
 DataParser(DataParser.TYPE_EXPLORER,
@@ -109,17 +117,15 @@ DataParser(DataParser.TYPE_EXPLORER,
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'ParentTxFrom', 'ParentTxTo',
             'ParentTxETH_Value', 'From', 'TxTo', 'ContractAddress', 'Value_IN(HT)', 'Value_OUT(HT)',
             None, 'Historical $Price/HT', 'Status', 'ErrCode', 'Type', 'PrivateNote'],
-           worksheet_name="HecoInfo",
+           worksheet_name=WORKSHEET_NAME,
            row_handler=parse_hecoinfo_internal)
-
-
 
 # Same header as Etherscan
 #DataParser(DataParser.TYPE_EXPLORER,
 #           "HecoInfo (HRC-20 Tokens)",
 #           ['Txhash', 'UnixTimestamp', 'DateTime', 'From', 'To', 'Value', 'ContractAddress',
 #            'TokenName', 'TokenSymbol'],
-#           worksheet_name="HecoInfo",
+#           worksheet_name=WORKSHEET_NAME,
 #           row_handler=parse_hecoinfo_tokens)
 
 # Same header as Etherscan
@@ -127,5 +133,5 @@ DataParser(DataParser.TYPE_EXPLORER,
 #           "HecoInfo (HRC-721 NFTs)",
 #           ['Txhash', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress', 'TokenId',
 #            'TokenName', 'TokenSymbol'],
-#           worksheet_name="HecoInfo",
+#           worksheet_name=WORKSHEET_NAME,
 #           row_handler=parse_hecoinfo_nfts)